chore: move test users to CLI command
