+++ database/seeders/DatabaseSeeder.php
@@ -2,7 +2,6 @@
 
 namespace Database\Seeders;
 
-use App\Models\SearchToken;
 use Illuminate\Database\Seeder;
 
 class DatabaseSeeder extends Seeder
@@ -14,10 +13,8 @@ class DatabaseSeeder extends Seeder
      */
     public function run()
     {
-        $this->call(AdminSeeder::class);
         $this->call(BookSeeder::class);
         $this->call(CrawlerSeeder::class);
-        $this->call(EditorsSeeder::class);
         $this->call(PeopleSeeder::class);
         $this->call(RolesSeeder::class);
         $this->call(SuttasSeeder::class);
